+++ geo/src/lib.rs
@@ -204,10 +204,10 @@
 //!     - Enables [network grid] support for the [`proj` crate]
 //!     - After enabling this feature, [further configuration][proj crate file download] is required to use the network grid.
 //!     - ☐ Disabled by default
-//! - `use-proj`:
+//! - `proj`:
 //!     - Enables coordinate conversion and transformation of `Point` geometries using the [`proj` crate]
 //!     - ☐ Disabled by default
-//! - `use-serde`:
+//! - `serde`:
 //!     - Allows geometry types to be serialized and deserialized with [Serde]
 //!     - ☐ Disabled by default
 //! - `multithreading`:
@@ -248,7 +248,7 @@
 //! [rhumb line]: https://en.wikipedia.org/wiki/Rhumb_line
 //! [Serde]: https://serde.rs/
 
-#[cfg(feature = "use-serde")]
+#[cfg(feature = "serde")]
 #[macro_use]
 extern crate serde;
 
@@ -390,6 +390,32 @@ impl_geo_num_for_int!(i32);
 impl_geo_num_for_int!(i64);
 impl_geo_num_for_int!(i128);
 impl_geo_num_for_int!(isize);
+
+// Some gymnastics to help migrate people off our old feature flag naming conventions
+#[allow(unused)]
+mod deprecated_feature_flags {
+    #[cfg_attr(
+        not(feature = "__allow_deprecated_features"),
+        deprecated(
+            since = "0.31.1",
+            note = "The `use-serde` feature has been renamed to simply `serde`. Use the `serde` feature instead."
+        )
+    )]
+    pub struct UseSerde;
+
+    #[cfg_attr(
+        not(feature = "__allow_deprecated_features"),
+        deprecated(
+            since = "0.31.1",
+            note = "The `use-proj` feature has been renamed to simply `proj`. Use the `proj` feature instead."
+        )
+    )]
+    pub struct UseProj;
+}
+#[cfg(feature = "use-proj")]
+pub use deprecated_feature_flags::UseProj;
+#[cfg(feature = "use-serde")]
+pub use deprecated_feature_flags::UseSerde;
 
 #[cfg(test)]
 mod tests {