+++ geo/src/types.rs
@@ -1,7 +1,7 @@
 use crate::{GeoFloat, Point};
 
 /// The result of trying to find the closest spot on an object to a point.
-#[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Closest<F: GeoFloat> {
     /// The point actually intersects with the object.